% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeiHMM.R
\name{plotMeiHMM}
\alias{plotMeiHMM}
\title{plotMeiHMM function}
\usage{
plotMeiHMM(MeiHMM.results)
}
\arguments{
\item{MeiHMM.results}{Returned list of function MeiHMM}
}
\description{
plotMeiHMM generate figures for MeiHMM results for visual inspection
}
