% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeiHMM.R
\name{MeiHMM}
\alias{MeiHMM}
\title{MeiHMM function}
\usage{
MeiHMM(snp.file, ancestry = c("tot"))
}
\arguments{
\item{snp.file}{Path to the input file}

\item{ancestry}{The ancestry groups, can be a list in c("eur", "amr", "eas", "sas", "afr"). Use "tot" to ignore ancestry.}
}
\value{
A list containing: segmentation, the segmentation results; positions, position of markers, af.doubled.allele, population allele frequency of two-copy alleles; oe.score, O/E ratio of hypothetical haplotypes
}
\description{
MeiHMM segment chromosome 21 into blocks of two or three haplotypes
}
